# net/http: use Copy in ServeContent if CopyN not needed