--- a/src/net/http/fs_test.go
+++ b/src/net/http/fs_test.go
@@ -928,6 +928,7 @@ func testServeContent(t *testing.T, mode testMode) {
 		wantContentType  string
 		wantContentRange string
 		wantStatus       int
+		wantContent      []byte
 	}
 	htmlModTime := mustStat(t, "testdata/index.html").ModTime()
 	tests := map[string]testCase{
@@ -1148,6 +1149,7 @@ func testServeContent(t *testing.T, mode testMode) {
 			serveContentType: "text/plain; charset=utf-8",
 			wantContentType:  "text/plain; charset=utf-8",
 			wantStatus:       StatusOK,
+			wantContent:      []byte("foobar"),
 		},
 		"do_not_use_writeTo_for_range_requests": {
 			content:          &panicOnWriterTo{ReadSeeker: strings.NewReader("foobar")},
@@ -1158,6 +1160,7 @@ func testServeContent(t *testing.T, mode testMode) {
 			wantContentType:  "text/plain; charset=utf-8",
 			wantContentRange: "bytes 0-4/6",
 			wantStatus:       StatusPartialContent,
+			wantContent:      []byte("fooba"),
 		},
 	}
 	for testName, tt := range tests {
@@ -1172,7 +1175,8 @@ func testServeContent(t *testing.T, mode testMode) {
 		} else {
 			content = tt.content
 		}
-		for _, method := range []string{"GET", "HEAD"} {
+		contentOut := &strings.Builder{}
+		for _, method := range []string{MethodGet, MethodHead} {
 			//restore content in case it is consumed by previous method
 			if content, ok := content.(*strings.Reader); ok {
 				content.Seek(0, io.SeekStart)
@@ -1198,7 +1202,8 @@ func testServeContent(t *testing.T, mode testMode) {
 			if err != nil {
 				t.Fatal(err)
 			}
-			io.Copy(io.Discard, res.Body)
+			contentOut.Reset()
+			io.Copy(contentOut, res.Body)
 			res.Body.Close()
 			if res.StatusCode != tt.wantStatus {
 				t.Errorf("test %q using %q: got status = %d; want %d", testName, method, res.StatusCode, tt.wantStatus)
@@ -1211,6 +1216,9 @@ func testServeContent(t *testing.T, mode testMode) {
 			}
 			if g, e := res.Header.Get("Last-Modified"), tt.wantLastMod; g != e {
 				t.Errorf("test %q using %q: got last-modified = %q, want %q", testName, method, g, e)
+			}
+			if g, e := contentOut.String(), tt.wantContent; e != nil && method == MethodGet && g != string(e) {
+				t.Errorf("test %q using %q: got unexpected content %q, want %q", testName, method, g, e)
 			}
 		}
 	}