@costela 2024-01-15
although the [excelent benchmark](https://github.com/golang/go/issues/61530#issuecomment-1646994568) by @AlexanderYastrebov seemed to indicate the underlying issue was solved, the [higher-level benchmark](https://github.com/golang/go/issues/61530#issuecomment-1646701847) by @mauri870 still shows a measurable slowdown (though not quite on the same level as originally observed; tested on current `master`):
```
goos: linux
goarch: amd64
pkg: go61530
cpu: 12th Gen Intel(R) Core(TM) i7-1250U
             │  old.bench  │             new.bench              │
             │   sec/op    │   sec/op     vs base               │
FileServe-12   17.27µ ± 6%   18.12µ ± 2%  +4.94% (p=0.007 n=10)

             │  old.bench   │               new.bench               │
             │     B/op     │     B/op      vs base                 │
FileServe-12   19.97Ki ± 0%   42.73Ki ± 0%  +114.01% (p=0.000 n=10)

             │ old.bench  │             new.bench              │
             │ allocs/op  │ allocs/op   vs base                │
FileServe-12   25.00 ± 0%   28.00 ± 0%  +12.00% (p=0.000 n=10)
```

Since the original [syntetic benchmarks](https://github.com/golang/go/pull/56480#issuecomment-1295451380) remain promising, this looks like another negative interaction with `os.File`.

I'll draft this PR while I investigate further. Sorry for the noise.