@costela 2024-01-15
TL;DR: benchmarks were skewed; this PR should cause no performance regressions :tada: :crossed_fingers: 

Longer:

The benchmark results above are skewed by the using `httptest.ResponseRecorder`, which causes a fallback to `os.genericWriteTo` for not being one of the supported targets for a `WriteTo` from an `os.File`. 

Old (`CopyN`)
![image](https://github.com/golang/go/assets/94699/7ad2ba33-e347-4553-aed7-63dfc8050012)

New (`Copy`)
![image](https://github.com/golang/go/assets/94699/8fb5d750-59e7-4611-a8cd-a9389c0df7e8)

The overhead is then caused by missing the [optimization in `copyBuffer` for `LimitReader`](https://github.com/golang/go/blob/7abeefd2b1a03932891e581f1f90656ffebebce4/src/io/io.go#L419). It pretty much disappears as soon as the the response size is bigger than the standard buffer size of 32k:
```
goos: linux
goarch: amd64
pkg: go61530
cpu: 12th Gen Intel(R) Core(TM) i7-1250U
             │  old.bench  │             new.bench              │
             │   sec/op    │   sec/op     vs base               │
FileServe-12   1.425m ± 2%   1.453m ± 5%  +1.95% (p=0.005 n=10)

             │  old.bench   │           new.bench            │
             │     B/op     │     B/op      vs base          │
FileServe-12   75.58Ki ± 3%   76.43Ki ± 3%  ~ (p=0.280 n=10)

             │ old.bench  │             new.bench              │
             │ allocs/op  │ allocs/op   vs base                │
FileServe-12   16.00 ± 0%   19.00 ± 0%  +18.75% (p=0.000 n=10)
```

Adding `ReadFrom` to the `http.ResponseRecorder` also negates the negative effects (I'd say this is an acceptable comparison with `http.response`, which also implements `ReaderFrom`):
```
goos: linux
goarch: amd64
pkg: go61530
cpu: 12th Gen Intel(R) Core(TM) i7-1250U
             │  old.bench  │           new.bench           │
             │   sec/op    │   sec/op     vs base          │
FileServe-12   1.018m ± 7%   1.032m ± 3%  ~ (p=0.165 n=10)

             │  old.bench   │           new.bench            │
             │     B/op     │     B/op      vs base          │
FileServe-12   32.45Ki ± 3%   31.10Ki ± 7%  ~ (p=0.165 n=10)

             │ old.bench  │             new.bench              │
             │ allocs/op  │ allocs/op   vs base                │
FileServe-12   15.00 ± 0%   18.00 ± 0%  +20.00% (p=0.000 n=10)
```

This means this PR shouldn't harm performance for the `io.File` case, and should enable `WriteTo` from sources that support if (e.g. [x/exp/mmap](https://github.com/golang/go/issues/56422#issue-1422973099), hopefully :crossed_fingers:)