@costela 2024-01-15
AFAICT the PR should actually _improve_ performance for the realistic `io.File` case because `ReadFrom` in `http.response` will have access to the `io.File` (instead of `io.LimitReader`)? :thinking: