@costela 2024-01-16
This still causes a regression in the `io.File` case:
![image](https://github.com/golang/go/assets/94699/ec9d33f6-0007-48ce-9b5d-585533a41cda)

1. `io.Copy` tries to use `WriteTo` from the `io.File` (not available if using `io.CopyN`), which fails because `http.response` doesn't implement `syscall.Conn`, so we use `genericWriteTo`
2. that calls a **second** round of `io.Copy`, now with a wrapped `io.File`, so we skip `WriteTo` and go straight to `http.response.ReadFrom`.
3. that then cannot use `sendFile`, because we got a wrapped file as source, so we're back to `genericReadFrom`
4. finally, a **third** round of `io.Copy`, which just falls back to buffered copy.