@costela 2024-01-16
With the latest commit the regression is solved. We now see a working `ReadFrom`+`sendFile in the `io.File` case:
![image](https://github.com/golang/go/assets/94699/05c2f3e2-e8ab-4566-bb7f-bad1c3194973)

However, since there are currently multiple proposals open around the intersection of `WriterTo` and `io.Copy` (e.g. #16474, #58331, #56742), it may be necessary to split the last commit into a proposal.

Alternatively, we could also add an internal interface that is only implemented by `io.File`, so that all type checks can be done against it. This would ensure this change has no side-effects outside of the stdlib. In its current form, the change has an unlikely impact, but could still theoretically accidentally match some interface in the wild.